refactor(prefixCommands): improve file cohesion
