+++ src/index.ts
@@ -1,7 +1,7 @@
 import "dotenv/config";
 import { Client, Collection, GatewayIntentBits, Partials } from "discord.js";
 import registerBotEvents from "./handler/on-start/bot-events.handler";
-import registerCommands from "./handler/on-start/commands.handler";
+import registerCommands from "./commands/commands.handler";
 
 const client = new Client({
   intents: [
